+++ package.json
@@ -35,10 +35,10 @@
 		"@codemirror/search": "^6.5.6",
 		"@codemirror/state": "^6.4.1",
 		"@codemirror/view": "^6.24.1",
-		"@dnd-kit/core": "^6.0.8",
-		"@dnd-kit/modifiers": "^6.0.1",
-		"@dnd-kit/sortable": "^7.0.2",
-		"@dnd-kit/utilities": "^3.2.1",
+		"@dnd-kit/core": "^6.1.0",
+		"@dnd-kit/modifiers": "^7.0.0",
+		"@dnd-kit/sortable": "^8.0.0",
+		"@dnd-kit/utilities": "^3.2.2",
 		"@lezer/highlight": "^1.2.0",
 		"@mantine/core": "^7.6.2",
 		"@mantine/hooks": "^7.6.2",
